Pass field to sort by in call to case notes api
For https://dsdmoj.atlassian.net/browse/NN-1706

When a case note merge happens a new note is created for
the occurrence. This results in notes relating to old
occurrence sometimes appearing before notes relating to newer and
potentially more relevant occurrences.

This is the first step to changing how case notes are displayed
in the interface.

Next step: Switch around creation date and occurrence date in the
interface.