Update interface to switch around case note occurrence and creation dates
For https://dsdmoj.atlassian.net/browse/NN-1706

This updates the CaseNoteListItem component to show the
occurrence date more prominently on the left side of the
list item (currently that's showing creation date) and to instead
show the creation date on the right hand-side of the item where
the occurence date lives right now. The labels are also updated.

I added a small change to the CaseNote data object used to
generate the testing snashop so that I could ensure the right
date was showing in the right place (as they are identical right now, it is impossible to tell the differece). It pulls the occurrence date
two hours earlier than the case note creation date.