Nn 2204 alert flags filter form (#331)
This adds more alert filters for the search page for an offender.
In order to avoid duplication of flags and alert codes and to
reduce the opportunity for error, this pulls all the alert
flags into a config JSON file from where both the flags and
the filters are powered.

There is an issue with this, however - there are two alert codes that
should result in the same flag and one filter which should correspond to
two alert codes. Currently there's just repetition.